part5: ex5.6
